examples/fastapi_sqlalchemy/requirements.txt
-e ../../
fastapi
starlette-graphene3
websockets
aiodataloader
